refactor(ActualTable): move component to its own directory
