+++ src/pages/reportes/actual/index.tsx
@@ -9,7 +9,7 @@ import BcaSelect from '~/components/input/BcaSelect/BcaSelect'
 import { useGetAllProjectsQuery } from '~redux/api/bca-backend/parametros/projectsSlice'
 import { useGetAllBudgetsByProjectAndLevelQuery } from '~redux/api/bca-backend/transacciones/budgetSlice'
 import { useGetAllLevelsQuery } from '~redux/api/bca-backend/reports/commonSlice'
-import ActualTable from '~components/reports/ActualTable'
+import ActualTable from '~/components/reports/ActualTable/ActualTable'
 import { useAppSelector } from '~redux/hooks'
 import EditToolbar from '~components/table/headers/toolbar'
 import { downloadExcelFile } from '~/utils/download'
