+++ src/pages/reportes/historic/index.tsx
@@ -13,7 +13,7 @@ import {
 } from '~redux/api/bca-backend/reports/commonSlice'
 import EditToolbar from '~components/table/headers/toolbar'
 import BcaDateTextField from '~/components/input/BcaDateTextField/BcaDateTextField'
-import ActualTable from '~components/reports/ActualTable'
+import ActualTable from '~/components/reports/ActualTable/ActualTable'
 import { useAppSelector } from '~redux/hooks'
 import { downloadExcelFile } from '~/utils/download'
 import { normalizeDate } from '~/utils/date'
